requirements.txt
Flask==2.3.3
pycryptodome==3.19.0
qrcode[pil]==7.4.2
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
